Working through stack overflow issues.
I was running into memory corruption problems during processing in the
console thread. Ultimately, the problem was just due to stack overflows,
but in the process of debugging it I ran into an article that points out
that the HAL doesn't integrate the newlib library very well with the
FreeRTOS memory allocation implementations for multi-threaded applications.

As a result of that finding, I made the changes recommended at this site:
http://nadler.com/embedded/newlibAndFreeRTOS.html

I don't know if the ODriveV2 software would have been affected by the
problems with FreeRTOS multi-threading and newlib because I am only ever
using printf from a single thread, but I will be leaving the changes for
now.